This function calculates postfix expression

This program computes the value by accepting the string input which is space-delimited arithmetic expression in postfix notation.

This program computes the value by accepting the string input which is space-delimited arithmetic expression in prefix notation.

check_valid_operators(operator)

check_operator(char)

check_digit(number)

perform_operation(z, x, y)

push(input_vector, values)

pop(input_vector)

calculate_postfix(input_string)

check_digit(number)

evaluate_prefix(expression)

Arguments

operator
A string with mathematical operators

char
This function checkd if the character value is

number
checks if input is a number

z
operand 2

x
operand 1

y
operator

input_vector
input vector

values
numeric values

input_string
A string with space-delimited integer arithmetic expression in postfix notation.

expression
A string with space-delimited integer arithmetic expression in prefix notation

Value

A integer value

A integer value

Examples

calculate_postfix("2 3 1 * + 9 - ")

#> [1] "The result is: -4"

expr <- "+ 9 * - 2 5 + 3 2"
evaluate_prefix(expr)

#> [1] "The result is: -6"