This function checks if the given string is a palindrome

Tests whether a string or sentence is a palindrome by removing special punctuation characters. A palindrome is equal to its reverse, for example "rotor" or "racecar".

is_palindrome(str)

Arguments

str
A string or full sentence

Value

TRUE if s is a palindrome and FALSE otherwise

Examples

is_palindrome("Was It A Rat I Saw?")

#> [1] TRUE