This function takes a vector and truns it into a data frame with multiple columns.

This function takes a vector and truns it into a data frame with multiple columns.

split_rows_to_columns(vector, number_rows, col_name_prefix = "name")

Arguments

vector
A vector

number_rows
Number of rows to split the vector into data frame

col_name_prefix
A prefix for column names for the data frame

Value

A data frame with multiple columns

Examples

split_rows_to_columns(mtcars$mpg, number_rows = 7, col_name_prefix = "ID List")

#>   ID List 1 ID List 2 ID List 3 ID List 4 ID List 5
#> 1      21.0      24.4      10.4      15.5      15.8
#> 2      21.0      22.8      10.4      15.2      19.7
#> 3      22.8      19.2      14.7      13.3      15.0
#> 4      21.4      17.8      32.4      19.2      21.4
#> 5      18.7      16.4      30.4      27.3        NA
#> 6      18.1      17.3      33.9      26.0        NA
#> 7      14.3      15.2      21.5      30.4        NA

split_rows_to_columns(mtcars$mpg, number_rows = 1, col_name_prefix = "ID List")

#>    ID List 1
#> 1       21.0
#> 2       21.0
#> 3       22.8
#> 4       21.4
#> 5       18.7
#> 6       18.1
#> 7       14.3
#> 8       24.4
#> 9       22.8
#> 10      19.2
#> 11      17.8
#> 12      16.4
#> 13      17.3
#> 14      15.2
#> 15      10.4
#> 16      10.4
#> 17      14.7
#> 18      32.4
#> 19      30.4
#> 20      33.9
#> 21      21.5
#> 22      15.5
#> 23      15.2
#> 24      13.3
#> 25      19.2
#> 26      27.3
#> 27      26.0
#> 28      30.4
#> 29      15.8
#> 30      19.7
#> 31      15.0
#> 32      21.4